pyproject.toml
[project]
name = "sbs-reportr"
version = "0.1.0"
description = "Backend API for the SBS reporting workflow."
readme = "PLAN.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi[standard]>=0.116.0",
  "pillow>=11.2.0",
  "pydantic>=2.12.0",
  "weasyprint>=66.0",
]

[dependency-groups]
dev = [
  "basedpyright>=1.31.5",
  "pytest>=8.4.2",
  "ruff>=0.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]
extend-exclude = [
  "src/frontend/node_modules",
  "src/frontend/dist",
  "src/frontend/dist-ssr",
  "src/frontend/coverage",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
